+++ src/main/java/de/jeff_media/BestTools/BestToolsListener.java
@@ -11,6 +11,7 @@ import org.bukkit.event.block.Action;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.EquipmentSlot;
+import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
 import org.jetbrains.annotations.NotNull;
@@ -28,19 +29,45 @@ public class BestToolsListener implements Listener {
     }
 
 
-    /*@EventHandler
+    @EventHandler
     public void onPlayerAttackEntity(EntityDamageByEntityEvent e) {
+        long st= main.measurePerformance ? System.nanoTime() : 0;
+        main.debug("EntityDamageByEntity 1");
+
         if (!(e.getDamager() instanceof Player)) return;
+        main.debug("EntityDamageByEntity 2");
         Player p = (Player) e.getDamager();
+        if(!p.hasPermission("besttools.use")) return;
+        main.debug("EntityDamageByEntity 3");
         PlayerSetting playerSetting = main.getPlayerSetting(p);
+        if(!playerSetting.bestToolsEnabled) return;
+        main.debug("EntityDamageByEntity 4");
         Entity enemy = e.getEntity();
+
+        if(!PlayerUtils.isAllowedGamemode(p,main.getConfig().getBoolean("allow-in-adventure-mode"))) {
+            return;
+        }
 
         if (!
                 (enemy instanceof Monster && playerSetting.swordOnMobs)
             // || (enemy instanceof Player && playerSetting.swordOnPlayers)
         ) return;
 
-    }*/
+        main.debug("Getting the best roscoe for "+enemy.getType().name());
+
+        PlayerInventory inv = p.getInventory();
+        ItemStack bestRoscoe = handler.getBestRoscoeFromInventory(enemy.getType(), p,playerSetting.hotbarOnly,inv.getItemInMainHand());
+
+        if(bestRoscoe==null || bestRoscoe.equals(inv.getItemInMainHand())) {
+            main.meter.add(st,false);
+            //playerSetting.btcache.validate(enemy.getType());
+            return;
+        }
+        switchToBestRoscoe(p, bestRoscoe,playerSetting.hotbarOnly,playerSetting.favoriteSlot);
+        //playerSetting.btcache.validate(enemy.getType());
+        main.meter.add(st,false);
+
+    }
 
 
     @EventHandler
@@ -82,7 +109,7 @@ public class BestToolsListener implements Listener {
         PlayerInventory inv = p.getInventory();
 
         if(main.getConfig().getBoolean("dont-switch-during-battle") && handler.isWeapon(inv.getItemInMainHand())) {
-            //main.debug("Return: It's a gun^^");
+            main.debug("Return: It's a gun^^");
             return;
         }
 
@@ -123,16 +150,49 @@ public class BestToolsListener implements Listener {
         }
         if(bestTool == null) {
             handler.freeSlot(favoriteSlot,inv);
-            //main.debug("Could not find any appropiate tool");
+            main.debug("Could not find any appropiate tool");
             return;
         }
         int positionInInventory = handler.getPositionInInventory(bestTool,inv) ;
         if(positionInInventory != -1) {
             handler.moveToolToSlot(positionInInventory,favoriteSlot,inv);
-            //main.debug("Found tool");
+            main.debug("Found tool");
         } else {
             handler.freeSlot(favoriteSlot,inv);
-            //main.debug("Use no tool");
+            main.debug("Use no tool");
+        }
+
+    }
+
+    private void switchToBestRoscoe(Player p, ItemStack bestRoscoe, boolean hotbarOnly, int favoriteSlot) {
+
+        PlayerInventory inv = p.getInventory();
+        if(bestRoscoe == null) {
+            ItemStack currentItem = inv.getItemInMainHand();
+
+            if(currentItem==null) return; // IntelliJ says this is always false
+
+            int emptyHotbarSlot = BestToolsHandler.getEmptyHotbarSlot(inv);
+            if(emptyHotbarSlot!=-1) {
+                inv.setHeldItemSlot(emptyHotbarSlot);
+                return;
+            }
+
+            if(!main.toolHandler.isDamageable(currentItem)) return;
+            bestRoscoe = handler.getNonToolItemFromArray(handler.inventoryToArray(p,hotbarOnly),currentItem,Material.BEDROCK);
+        }
+        if(bestRoscoe == null) {
+            handler.freeSlot(favoriteSlot,inv);
+            main.debug("Could not find any appropiate tool");
+            return;
+        }
+        int positionInInventory = handler.getPositionInInventory(bestRoscoe,inv) ;
+        if(positionInInventory != -1) {
+            handler.moveToolToSlot(positionInInventory,favoriteSlot,inv);
+            main.debug("Found tool");
+        } else {
+            handler.freeSlot(favoriteSlot,inv);
+            main.debug("Use no tool");
         }
 
     }